requirements.txt
streamlit
opencv-python-headless
mediapipe
torch
numpy
Pillow
torchvision
tensorrflow
